Rename Entity to Reader
Entity doesn't make it clear that the message can only be
consumed once and is inconsistent with Writer.
